requirements.txt
flask 
flask-login
flask-wtf
bcrypt
libsql-experimental
python-dotenv